changed default search results to display location
Instead of relying on blacklight_config to display out the index
fields, the default index page is now a bespoke page. It displays
URLs for online results, and location and call number for other items.

Adding the relevant fields (location_display and others) necessitated
changes to the solr schema and solr config.