added generation of locations_fixture, added hours formatting and finding
lib/generate_location_fixture.rb now pulls from the location page to
generate a fixtures file that has all the location information.

it is imperfect, but a good start.

this is enhanced with syncing features on location

also, have added the ability to find a range of dates easily,
as well as is_open? for a given time.