fixed folders, sms, emailing
had to copy a lot of files locally for this: the default
forms are still referring to email_sms_record, which is
both deprecated, and doesn't work right.

Added sms multiple items capabiltities (although should
put in a warning about data charges, woe to the user
who sms's 10 of them off)

Removed the custom folder js we were using, and instead
fixed up the  checking for in folder mechanism (which
was having trouble with the fact that it was an array
of ids.

newbooks-79 and newbooks-8 knocked out here.